feat:vscode: add ability to select labs
currently the west custom build requires main.c to be focused
in order to determine which lab to compile. This can be error prone
during training, and is often overlooked.

a way around this would be select the lab as an input option when
running the task

tested with labs, able to select and compile different labs without
having a specific file in focus